refactor(config): Move sudo helpers to platform-specific files
Move getSudoUserInfo, getHomeDir, ChownToUser, and
ChownToUserRecursive to utils_unix.go with go:build unix
constraint. Create utils_windows.go with stubbed implementations
for Windows compatibility, maintaining the same interface.
