+++ config/config_file.go
@@ -27,12 +27,9 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"log"
 	"os"
-	"os/user"
 	"path/filepath"
 	"runtime"
-	"strconv"
 	"syscall"
 )
 
@@ -44,114 +41,6 @@ const (
 	APP_DATA            = "AppData"
 	LOCAL_APP_DATA      = "LocalAppData"
 )
-
-// sudoUserInfo contains information about the original user when running under sudo.
-type sudoUserInfo struct {
-	HomeDir string
-	UID     int
-	GID     int
-}
-
-// getSudoUserInfo returns information about the original user when running under sudo.
-// Returns nil if not running under sudo or if the original user cannot be determined.
-func getSudoUserInfo() *sudoUserInfo {
-	sudoUser := os.Getenv("SUDO_USER")
-	if sudoUser == "" {
-		return nil
-	}
-
-	// Try to get UID/GID from environment first (more reliable)
-	uidStr := os.Getenv("SUDO_UID")
-	gidStr := os.Getenv("SUDO_GID")
-
-	var uid, gid int
-	var homeDir string
-
-	// Look up the user to get their home directory
-	if u, err := user.Lookup(sudoUser); err == nil {
-		homeDir = u.HomeDir
-		// Use looked up UID/GID as fallback
-		if uidStr == "" {
-			uidStr = u.Uid
-		}
-		if gidStr == "" {
-			gidStr = u.Gid
-		}
-	}
-
-	if homeDir == "" {
-		return nil
-	}
-
-	// Parse UID
-	if uidStr != "" {
-		if parsedUID, err := strconv.Atoi(uidStr); err == nil {
-			uid = parsedUID
-		}
-	}
-
-	// Parse GID
-	if gidStr != "" {
-		if parsedGID, err := strconv.Atoi(gidStr); err == nil {
-			gid = parsedGID
-		}
-	}
-
-	return &sudoUserInfo{
-		HomeDir: homeDir,
-		UID:     uid,
-		GID:     gid,
-	}
-}
-
-// getHomeDir returns the appropriate home directory for kraftkit configuration.
-// When running under sudo, this returns the original user's home directory
-// to maintain consistent config paths between privileged and unprivileged commands.
-func getHomeDir() string {
-	if info := getSudoUserInfo(); info != nil {
-		return info.HomeDir
-	}
-	homeDir, err := os.UserHomeDir()
-	if err != nil {
-		log.Printf("warning: could not determine home directory: %v", err)
-	}
-	return homeDir
-}
-
-// ChownToUser changes ownership of a path to the original user when running under sudo.
-// This ensures that files created by sudo commands are accessible by the regular user.
-// If not running under sudo, this is a no-op.
-func ChownToUser(path string) error {
-	info := getSudoUserInfo()
-	if info == nil || info.UID == 0 {
-		// Not running under sudo or couldn't determine original user
-		return nil
-	}
-
-	return os.Chown(path, info.UID, info.GID)
-}
-
-// ChownToUserRecursive changes ownership of a path and all its contents to the original user.
-// This is useful for directories created by sudo commands.
-// Only files/dirs owned by root are chowned to avoid unnecessary syscalls.
-func ChownToUserRecursive(path string) error {
-	info := getSudoUserInfo()
-	if info == nil || info.UID == 0 {
-		return nil
-	}
-
-	return filepath.Walk(path, func(name string, fileInfo os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-		if stat, ok := fileInfo.Sys().(*syscall.Stat_t); ok {
-			if stat.Uid != 0 {
-				return nil
-			}
-		}
-		return os.Chown(name, info.UID, info.GID)
-	})
-}
 
 // Config path precedence
 // 1. KRAFTKIT_CONFIG_DIR
